Move CodeExitError to remocommand package and remote unnecessary copies
