+++ pkg/kubelet/remote/BUILD
@@ -20,10 +20,10 @@ go_library(
         "//pkg/kubelet/apis/cri:go_default_library",
         "//pkg/kubelet/apis/cri/v1alpha1:go_default_library",
         "//pkg/kubelet/util:go_default_library",
-        "//pkg/util/exec:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/golang.org/x/net/context:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
+        "//vendor/k8s.io/client-go/tools/remotecommand:go_default_library",
     ],
 )
