+++ staging/src/k8s.io/client-go/tools/remotecommand/BUILD
@@ -27,6 +27,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "doc.go",
+        "errors.go",
         "errorstream.go",
         "remotecommand.go",
         "resize.go",
@@ -46,6 +47,5 @@ go_library(
         "//vendor/k8s.io/client-go/pkg/api/v1:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
         "//vendor/k8s.io/client-go/transport:go_default_library",
-        "//vendor/k8s.io/client-go/util/exec:go_default_library",
     ],
 )
