--- staging/src/k8s.io/client-go/util/exec/BUILD
@@ -1,14 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-licenses(["notice"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["exec.go"],
-    tags = ["automanaged"],
-)
